// schema/v/code/library.d.ts
type sql = string;
// ...
export type dbname = string;
//
//The table in the named dbname where the data will be saved 
export type ename = string;
//
//The number of similar records being saved
type alias = Array<number>;
//
//COlumn name in the database table to be saved 
type cname = string;
//
//The primary key value which is modeled as a number beause we 
//often use it for specifying the ofset of an sql result.
export type pk= number;
//
//The location of a td in a crud table. This is important for saving and 
//restoring a crud td. 
export type position =[rowIndex,cellIndex];
//
//The actual value being sent to the server 
export type basic_value = boolean|number|string|null;
// ...
export interface Imerge{
  dbname:string,
  ename:string,
  members:sql,
  minors?:sql,
  principal?:number  
}

type interventions = Array<intervention>;

interface intervention{
    cname:cname, 
    value:basic_value
}

type conflicts = Array<conflict>;

interface conflict{
    cname:cname,
    values: Array<basic_value>
}

//A pointer is a foreign key column of a contributor whose integrity would
//be violated if there is an attempt to delete a reference record. The pointer
//would need re-direction before being deleted.
export type pointer = {
    //
    //Name of the column/field
    cname: cname,
    //
    //Name of the table containing the column
    ename: ename,
    //
    //Name of the database in which the table is contained. 
    dbname: dbname,
    //
    //Is the foreign key a cross member or not. This is important 
    //for addressing possibiliy for cyclic referencing
    is_cross_member:boolean,
}; 

//Indices of the pointer's away table that can 
//be violated by re-direction
export type index = {
    //
    //Index name (for reporting purposes)
    ixname:string;
    //
    //An sql that generates a merging singature bases on the
    //violaters, i.e., the columns of an index needed to determine
    //integrity violaters. The sql has the shape:
    //Array<{signature}>
    signatures:Array<{signature:string}>, 
    //
    //The sql that is to be constrained by a specific to generate
    //the pointer members that need to be merged. he sql has the shape:-
    //Array<{signature, member}
    members:sql
}